--- a/.gitignore
+++ b/.gitignore
@@ -11,6 +11,7 @@ node_modules
 # testing
 /coverage
 .cache
+.eslintcache
 
 # code-gen
 /packages/icons/src/generated