--- a/.prettierignore
+++ b/.prettierignore
@@ -1,4 +1,4 @@
-packages/**/dist/
+packages/**/lib/
 packages/icons/src/generated/
 packages/www/.cache
 packages/www/public