--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -27,6 +27,7 @@
     "labelledby",
     "nomarker",
     "polarbrightness",
+    "popperjs",
     "spacebar",
     "stylelint",
     "testid",