--- a/jest.setup.js
+++ b/jest.setup.js
@@ -24,13 +24,14 @@
 
  */
 
-const Adapter = require('enzyme-adapter-react-16')
-const Enzyme = require('enzyme')
+import Adapter from 'enzyme-adapter-react-16'
+import { configure } from 'enzyme'
+
 require('@testing-library/jest-dom/extend-expect')
 require('jest-canvas-mock')
 require('jest-styled-components')
 
-Enzyme.configure({ adapter: new Adapter() })
+configure({ adapter: new Adapter() })
 
 const observeMock = function (cb, config) {
   this.observeCallback = cb