--- a/packages/components/src/Avatar/Avatar.tsx
+++ b/packages/components/src/Avatar/Avatar.tsx
@@ -59,7 +59,7 @@ export interface AvatarProps extends SpaceProps {
   size?: AvatarSizes | string
 }
 
-/* eslint-disable sort-keys */
+/* eslint-disable sort-keys,sort-keys-fix/sort-keys-fix */
 const size = variant({
   prop: 'size',
   variants: {