--- a/packages/components/src/Avatar/AvatarIcon.tsx
+++ b/packages/components/src/Avatar/AvatarIcon.tsx
@@ -48,7 +48,7 @@ export interface AvatarIconProps extends AvatarProps {
   bg?: string
 }
 
-/* eslint-disable sort-keys,sort-keys-fix/sort-keys-fix */
+/* eslint-disable sort-keys-fix/sort-keys-fix */
 const size = variant({
   prop: 'size',
   variants: {