--- a/packages/components/src/Calendar/Calendar.tsx
+++ b/packages/components/src/Calendar/Calendar.tsx
@@ -26,7 +26,6 @@
 import React, { FC } from 'react'
 import DayPicker, { RangeModifier } from 'react-day-picker'
 import 'react-day-picker/lib/style.css'
-import MomentLocaleUtils from 'react-day-picker/moment'
 import styled from 'styled-components'
 import has from 'lodash/has'
 import { mix } from 'polished'
@@ -36,24 +35,6 @@ import { inputTextFocus } from '../Form/Inputs/InputText'
 import { CalendarSize, calendarSize, calendarSpacing } from './calendar-size'
 import { CalendarContext } from './CalendarContext'
 import { CalendarNav } from './CalendarNav'
-
-// Import required data for all supported locales
-import 'moment/locale/ar'
-import 'moment/locale/de'
-import 'moment/locale/es'
-import 'moment/locale/fr'
-import 'moment/locale/it'
-import 'moment/locale/ja'
-import 'moment/locale/ko'
-import 'moment/locale/nl'
-import 'moment/locale/pl'
-import 'moment/locale/pt'
-import 'moment/locale/pt-br'
-import 'moment/locale/ru'
-import 'moment/locale/sv'
-import 'moment/locale/tr'
-import 'moment/locale/zh-cn'
-import 'moment/locale/zh-tw'
 
 interface CalendarProps {
   locale?: LocaleCodes
@@ -103,7 +84,6 @@ const InternalCalendar: FC<CalendarProps> = ({
       <DayPicker
         className={`${renderDateRange && 'render-date-range'} ${className}`}
         selectedDays={selectedDates}
-        localeUtils={MomentLocaleUtils}
         locale={locale}
         month={viewMonth}
         showOutsideDays={true}