--- a/packages/components/src/Form/Inputs/InlineTextArea/InlineTextArea.tsx
+++ b/packages/components/src/Form/Inputs/InlineTextArea/InlineTextArea.tsx
@@ -25,7 +25,7 @@
  */
 
 import pick from 'lodash/pick'
-import React, { forwardRef, Ref } from 'react'
+import React, { ChangeEvent, forwardRef, Ref, useState } from 'react'
 import isFunction from 'lodash/isFunction'
 import styled from 'styled-components'
 import {
@@ -58,13 +58,11 @@ export const InlineTextAreaLayout = forwardRef(
     }: InlineTextAreaProps,
     ref: Ref<HTMLTextAreaElement>
   ) => {
-    const [value, setValueChange] = React.useState(valueProp || '')
+    const [value, setValueChange] = useState(valueProp || '')
 
     const displayValue = isFunction(onChange) ? valueProp : value
 
-    const handleValueChange = (
-      event: React.ChangeEvent<HTMLTextAreaElement>
-    ) => {
+    const handleValueChange = (event: ChangeEvent<HTMLTextAreaElement>) => {
       setValueChange(event.currentTarget.value)
     }