--- a/packages/design-tokens/src/system/color/index.ts
+++ b/packages/design-tokens/src/system/color/index.ts
@@ -24,6 +24,6 @@
 
  */
 
-export { ColorProps } from 'styled-system'
+export type { ColorProps } from 'styled-system'
 export { Palette } from './palette'
 export { SemanticColor, SemanticColors } from './semantic_colors'