--- a/packages/design-tokens/src/tokens/color/palette.ts
+++ b/packages/design-tokens/src/tokens/color/palette.ts
@@ -26,7 +26,7 @@
 
 import { Palette } from '../../system'
 
-/* eslint-disable sort-keys,sort-keys-fix/sort-keys-fix */
+/* eslint-disable sort-keys-fix/sort-keys-fix */
 
 export const palette: Palette = {
   textInverted: '#FFFFFF',