--- a/packages/design-tokens/src/tokens/typography/font_sizes.ts
+++ b/packages/design-tokens/src/tokens/typography/font_sizes.ts
@@ -27,7 +27,7 @@
 import { rem } from 'polished'
 import { FontSizeRamp } from '../../system'
 
-/* eslint-disable sort-keys */
+/* eslint-disable sort-keys,sort-keys-fix/sort-keys-fix */
 export const fontSizes: FontSizeRamp = {
   xxxxlarge: rem(46),
   xxxlarge: rem(36),