--- a/packages/design-tokens/src/tokens/typography/font_weights.ts
+++ b/packages/design-tokens/src/tokens/typography/font_weights.ts
@@ -26,7 +26,7 @@
 
 import { FontWeightRamp } from '../../system'
 
-/* eslint-disable sort-keys,sort-keys-fix/sort-keys-fix */
+/* eslint-disable sort-keys-fix/sort-keys-fix */
 
 export const fontWeights: FontWeightRamp = {
   light: 300,