### :sparkles: Changes

This PR switches the monorepo to use Babel (rather than `tsc` / `ttsc` to build the JS artifacts. Additionally it enables for internal use of the packages to no longer require pre-compilation meaning that the playground and Gatsby can show live-edits to any package leveraged within those.

| Task  | Before | After |  |
|-------|-------:|------:|---------|
| `prepare` | 160.5s | 15.1s | 🎉  assuming that `node_modules` is already in normalized state (no calls to NPM req'd) |
| `build` | 57.7s | 51.3s | 🤷‍♂️ |
| `lint:es` | 102.1s | 80.7s | Slower than I hoped due to cache-writing and addition of `*.mdx` files. However no longer requires build before run. This means lint & build can be run concurrently for faster CI runs. | 
| `lint:es` (cache-hit) | n/a |1.5s | 🎉🎉🎉 ESLint leverages caching so that secondary runs only run against files that have changed. |
| `lint:mdx` | 8.8s | n/a | 🎉  In new build system `*.mdx` is linted as part of larger ESLint run |  
| `lint:ts` | 47.5s | 28.6s | 🎉 |
| `test` | 195.6s | 343.4s | Needs further investigation | 

#### Reviewer notes

`packages/playground/*` & `packages/www/*` weren't being fully linted before this PR so there's a lot of noise in this PR to straighten that out (mostly auto-fixed object order issues)

#### Follow-ups

This items will be addressed in follow-up tickets. This list should be checked-off once those tickets have been created.

- [x] Correct & re-enable tests on Confirm & Prompt
- [x] Fix former usage of `moment.js`
- [x] Pre/post-build steps to rearrange source for publishing (remove `/lib` from artifact paths)
- [x] Investigate `jest` slow-down
- [ ] Research why `lint:es - "import/default": "off"` is so problematic now (perhaps it was never _actually_ running)


### :white_check_mark: Requirements

- [x] Includes test coverage for all changes
- [x] Build and tests are passing
- [x] Update documentation
- [x] Updated CHANGELOG
- [x] Checked for i18n impacts
- [x] Checked for a11y impacts
- [ ] PR is ideally < ~400LOC

### :camera: Screenshots